Adds a new handler that allows you to get the players in a given manager's team for the current game week and get any recommended transfers for the next game week.
Refactored the linear programming code into a builder class that is shared by multiple handlers. Improved the weighting for players by incorporating their current form and the difficulty of their upcoming fixtures.

+++ src/main/kotlin/com/gavinflood/fpl/api/http/response/ManagerPicksResponseDTOs.kt
@@ -1,7 +1,7 @@
 package com.gavinflood.fpl.api.http.response
 
 data class ManagerPicksResponse(
-    // TODO: val active_chip:
+    //val active_chip: String,
     val automatic_subs: List<ManagerPicksResponseAutomaticSub>,
     val entry_history: ManagerPicksResponseHistory,
     val picks: List<ManagerPicksResponsePick>
@@ -29,7 +29,7 @@ data class ManagerPicksResponseHistory(
 )
 
 data class ManagerPicksResponsePick(
-    val element: Long,
+    val element: Int,
     val position: Int,
     val multiplier: Int,
     val is_captain: Boolean,
